made changes

+++ backend/database/migrations/2026_06_03_085339_create_server_metrics_table.php
@@ -20,8 +20,10 @@ return new class extends Migration
             $table->float('memory_usage')->nullable();
             $table->float('disk_usage')->nullable();
             $table->integer('process_count')->default(0);
-            $table->float('network_in')->default(0);
-            $table->float('network_out')->default(0);
+            
+            $table->unsignedBigInteger('network_in')->default(0);
+            $table->unsignedBigInteger('network_out')->default(0);
+            
             $table->timestamp('recorded_at')->useCurrent();
             $table->timestamps();
             $table->index(['server_id', 'recorded_at']);
